change column name student and staff table

+++ model/activity.go
@@ -11,7 +11,7 @@ type Activity struct {
 	Organizer     string       `json:"organizer"`
 	Category      string       `json:"category"`
 	Location      string       `json:"location"`
-	Semester      uint8        `json:"semester"`
+	Semester      uint16       `json:"semester"`
 	CompetencesID []uint16     `json:"competences_id,omitempty"` // For request
 	Competences   []Competence `json:"competences,omitempty"`    // For response
 	Attendees     []Student    `json:"attendees,omitempty"`
@@ -19,7 +19,7 @@ type Activity struct {
 }
 
 // NewActivity creates new activity struct
-func NewActivity(activityID uint32, activityName, description, date, time, creator, organizer, category, location string, semester uint8, studentSite bool) *Activity {
+func NewActivity(activityID uint32, activityName, description, date, time, creator, organizer, category, location string, semester uint16, studentSite bool) *Activity {
 	return &Activity{
 		ActivityID:   activityID,
 		ActivityName: activityName,
